Refactor: util 함수로 분리 #84
